+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- Fix bug where setting the place to none was not possible
 - Slightly improve predictions for vehicle and place when starting a recording or adding a new
   activity